+++ backend/core/src/listings/dto/listing.dto.ts
@@ -857,7 +857,7 @@ export class ListingFilterParams extends BaseFilter {
   })
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumberString({}, { groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.ami]?: number;
+  [ListingFilterKeys.minAmiPercentage]?: number;
 
   @Expose()
   @ApiProperty({
@@ -935,7 +935,10 @@ export class ListingsRetrieveQueryParams {
 // Fields for the Availability and AMI filters are determined based on the value
 // of the filter or by checking multiple columns. Since we can't specify a single
 // field the filters correspond to, we remove them from the filterTypeToFieldMap.
-type keysWithMappedField = Exclude<keyof typeof ListingFilterKeys, "ami" | "availability">
+type keysWithMappedField = Exclude<
+  keyof typeof ListingFilterKeys,
+  "minAmiPercentage" | "availability"
+>
 
 // Using a record lets us enforce that all types are handled in addFilter
 export const filterTypeToFieldMap: Record<keysWithMappedField, string> = {