+++ backend/core/src/listings/types/listing-filter-keys-enum.ts
@@ -8,7 +8,7 @@ export enum ListingFilterKeys {
   seniorHousing = "seniorHousing",
   minRent = "minRent",
   maxRent = "maxRent",
-  ami = "ami",
+  minAmiPercentage = "minAmiPercentage",
   leasingAgents = "leasingAgents",
 }