+++ backend/core/src/shared/filter/custom_filters.ts
@@ -67,7 +67,7 @@ export function addAvailabilityQuery(
   }
 }
 
-export function addAmiPercentageFilter(
+export function addMinAmiPercentageFilter(
   qb: WhereExpression,
   filterValue: number,
   includeNulls?: boolean