+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file. The format 
   - "About" page ([#589](https://github.com/CityOfDetroit/bloom/pull/589))
   - RTL support ([#627](https://github.com/CityOfDetroit/bloom/pull/627))
   - Add senior housing filtering to the filter modal ([#631](https://github.com/CityOfDetroit/bloom/pull/631))
+  - Add AMI filter to frontend ([\$645](https://github.com/CityOfDetroit/bloom/pull/645))
 
 ## Detroit Team M11